--- a/code/go/0chain.net/smartcontract/common/wallet.go
+++ b/code/go/0chain.net/smartcontract/common/wallet.go
@@ -1,28 +1,25 @@
 package common
 
 import (
+	"errors"
+
 	"0chain.net/chaincore/config"
-	"0chain.net/core/common"
 	"0chain.net/core/encryption"
-	"github.com/0chain/common/core/logging"
 )
 
-//ValidateDelegateWallet - Protects against using the provider's clientID (operational wallet ID) as DelegateWalletID. Checks that clientID and delegateWalletID are not the same
-func ValidateDelegateWallet(publicKey, delegateWalletID string) *common.Error {
+// ValidateDelegateWallet - Protects against using the provider's clientID (operational wallet ID) as DelegateWalletID. Checks that clientID and delegateWalletID are not the same
+func ValidateDelegateWallet(publicKey, delegateWalletID string) error {
 	if config.Development() {
 		return nil
 	}
 
 	operationalWalletID, err := encryption.GetClientIDFromPublicKey(publicKey)
 	if err != nil {
-		return common.NewError("add_sharder",
-			"couldn't decode publick key to compare to delegate wallet")
+		return errors.New("could not decode public key to compare to delegate wallet")
 	}
 
 	if operationalWalletID == delegateWalletID {
-		logging.Logger.Error("couldn't use the same wallet as both operational and delegate")
-		return common.NewError("add_sharder",
-			"couldn't use the same wallet as both operational and delegate")
+		return errors.New("could not use the same wallet as both operational and delegate")
 	}
 
 	return nil