--- a/code/go/0chain.net/smartcontract/dbs/event/eventdb.go
+++ b/code/go/0chain.net/smartcontract/dbs/event/eventdb.go
@@ -103,6 +103,9 @@ func (edb *EventDb) PageLimit() int64 {
 }
 
 func (edb *EventDb) Debug() bool {
+	if edb == nil {
+		return false
+	}
 	return edb.settings.Debug
 }