--- a/code/go/0chain.net/smartcontract/minersc/sharder.go
+++ b/code/go/0chain.net/smartcontract/minersc/sharder.go
@@ -124,6 +124,7 @@ func (msc *MinerSmartContract) AddSharder(
 	}
 
 	newSharder.NodeType = NodeTypeSharder // set node type
+	newSharder.ProviderType = spenum.Sharder
 	exist, err := msc.getSharderNode(newSharder.ID, balances)
 	if err != nil && err != util.ErrValueNotPresent {
 		return "", common.NewErrorf("add_sharder", "unexpected error: %v", err)