--- a/code/go/0chain.net/smartcontract/storagesc/benchmark_setup.go
+++ b/code/go/0chain.net/smartcontract/storagesc/benchmark_setup.go
@@ -456,6 +456,9 @@ func AddMockBlobbers(
 }
 
 func addMockBlobberSnapshots(blobber event.Blobber, edb *event.EventDb) {
+	if edb == nil {
+		return
+	}
 	var mockChallengesPassed = viper.GetUint64(sc.EventDbAggregatePeriod)
 	var mockChallengesCompleted = viper.GetUint64(sc.EventDbAggregatePeriod) + 1
 	const mockInactiveRounds = 17
@@ -486,6 +489,9 @@ func addMockBlobberSnapshots(blobber event.Blobber, edb *event.EventDb) {
 }
 
 func AddMockSnapshots(edb *event.EventDb) {
+	if edb == nil {
+		return
+	}
 	var snapshots []event.Snapshot
 	for i := 1; i < viper.GetInt(sc.NumBlocks); i += viper.GetInt(sc.EventDbAggregatePeriod) {
 		snapshot := event.Snapshot{