## Fixes

- closes #2001
- closes #2003
- fix #2002 partially
- fixes `miner` benchmark

## Changes
- Optimize smart contracts in minersc, including `add_miner`, `add_sharder`, `miner_health_check`, `sharder_health_check`, `sharder_keep`, etc.

```sh
benchmark_1  |
benchmark_1  | Timings
benchmark_1  |
benchmark_1  | Results
benchmark_1  |
benchmark_1  | benchmark suite miner
benchmark_1  | miner.addToDelegatePool,0.902857ms OK
benchmark_1  | miner.add_miner,1.190019ms OK
benchmark_1  | miner.add_sharder,1.072860ms OK
benchmark_1  | miner.collect_reward,1.048000ms OK
benchmark_1  | miner.contributeMpk,16.040000ms OK
benchmark_1  | miner.deleteFromDelegatePool,0.758598ms OK
benchmark_1  | miner.delete_miner,2.459514ms OK
benchmark_1  | miner.delete_sharder,1.084026ms OK
benchmark_1  | miner.miner_health_check,0.732212ms OK
benchmark_1  | miner.payFees,6.889535ms OK
benchmark_1  | miner.sharder_health_check,0.623014ms OK
benchmark_1  | miner.sharder_keep,0.870425ms OK
benchmark_1  | miner.shareSignsOrShares,4.597701ms OK
benchmark_1  | miner.update_globals,1.703808ms OK
benchmark_1  | miner.update_globals,0.728116ms OK
benchmark_1  | miner.update_miner_settings,0.675243ms OK
benchmark_1  | miner.update_settings,0.572633ms OK
benchmark_1  | miner.update_sharder_settings,0.598232ms OK
benchmarks_benchmark_1 exited with code 0
```
## TODO:
- [x]  Implement duplicate host and port checking 
- [x] Unit tests

## Need to be mentioned in CHANGELOG.md?

## Tests
Tasks to complete before merging PR:
- [ ]  Ensure system tests are passing. If not [Run them manually](https://github.com/0chain/0chain/actions/workflows/system_tests.yml) to check for any regressions :clipboard:
- [ ]  Do any new system tests need added to test this change? do any existing system tests need updated? If so create a PR at [0chain/system_test](https://github.com/0chain/system_test)
- [ ]  Merge your system tests PR to master AFTER merging this PR

## Associated PRs (Link as appropriate):
- blobber:
- gosdk:
- system_test:
- zboxcli:
- zwalletcli:
- Other: ...